TransferEvaluation_Django-master/requirements.txt
asgiref==3.2.10
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
et-xmlfile==1.0.1
jdcal==1.4.1
lml==0.0.9
openpyxl==3.0.5
pyexcel-io==0.6.2
pytz==2020.1
sqlparse==0.4.1
